+++ examples/react/cypress/components/hooks/useNodes.cy.tsx
@@ -15,8 +15,10 @@ const initialNodes: Node[] = nodes.map((n) => ({
 
 const expectedNodes: Node[] = initialNodes.map((n) => ({
   ...n,
-  positionAbsolute: n.position,
-  ...nodeDimensions,
+  computed: {
+    positionAbsolute: n.position,
+    ...nodeDimensions,
+  },
 }));
 
 describe('useNodes.cy.tsx', () => {
@@ -29,7 +31,6 @@ describe('useNodes.cy.tsx', () => {
       </ReactFlow>
     );
 
-    cy.get('@onChangeSpy').should('have.been.calledWith', []);
     cy.get('@onChangeSpy').should('have.been.calledWith', expectedNodes);
   });