+++ examples/react/src/examples/UseNodesData/TextNode.tsx
@@ -1,16 +1,21 @@
-import { memo, ChangeEventHandler } from 'react';
+import { memo, useState } from 'react';
 import { Position, NodeProps, Handle, useReactFlow } from '@xyflow/react';
 
 function TextNode({ id, data }: NodeProps) {
   const { updateNodeData } = useReactFlow();
-
-  const onChange: ChangeEventHandler<HTMLInputElement> = (evt) => updateNodeData(id, { text: evt.target.value });
+  const [text, setText] = useState(data.text);
+  const updateText = (text: string) => {
+    // avoid jumping caret with a synchronous update
+    setText(text);
+    // update actual node data
+    updateNodeData(id, { text });
+  };
 
   return (
     <div style={{ background: '#eee', color: '#222', padding: 10, fontSize: 12, borderRadius: 10 }}>
       <div>node {id}</div>
       <div>
-        <input onChange={onChange} value={data.text} />
+        <input onChange={(evt) => updateText(evt.target.value)} value={text} />
       </div>
       <Handle type="source" position={Position.Right} />
     </div>