+++ packages/react/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @xyflow/react
 
+## 12.0.0-next.6
+
+### Minor changes
+
+- fix `deleteElements`
+- refactor internal `applyChanges`
+- batch `setNodes` and `setEdges` from `useReactFlow`
+- add `aria-label` prop for `<Controls />`
+
 ## 12.0.0-next.5
 
 ### Minor changes